move subscription track to sperated file; fix incremental scan when resuming
