+++ components/backup-stream/src/lib.rs
@@ -10,6 +10,7 @@ pub mod metadata;
 mod metrics;
 pub mod observer;
 mod router;
+mod subscription_track;
 mod utils;
 
 pub use endpoint::{Endpoint, Task};
